+++ c3smembership/accountants_views.py
@@ -17,6 +17,7 @@ from datetime import (
     datetime,
     date,
 )
+from decimal import Decimal
 from types import NoneType
 
 import deform
@@ -238,6 +239,63 @@ def switch_pay(request):
         )
 
 
+def get_member_details(request, member):
+    """
+    Gets the member details.
+    """
+    invoices15 = Dues15Invoice.get_by_membership_no(member.membership_number)
+    invoices16 = Dues16Invoice.get_by_membership_no(member.membership_number)
+    invoices17 = Dues17Invoice.get_by_membership_no(member.membership_number)
+    shares = request.registry.share_information.get_member_shares(
+        member.membership_number)
+
+    return {
+        'today': date.today().strftime('%Y-%m-%d'),
+        'D': Decimal,
+        'member': member,
+        'shares': shares,
+        'invoices15': invoices15,
+        'invoices16': invoices16,
+        'invoices17': invoices17,
+    }
+
+
+@view_config(renderer='templates/detail.pt',
+             permission='manage',
+             route_name='member_details')
+def member_details(request):
+    """
+    This view lets accountants view member details:
+
+    - has their signature arrived?
+    - how about the payment?
+
+    Mostly all the info about an application or membership
+    in the database can be seen here.
+    """
+    logged_in = authenticated_userid(request)
+    membership_number = request.matchdict['membership_number']
+    LOG.info(
+        'member details of membership number %s checked by %s',
+        membership_number,
+        logged_in)
+
+    member_information = request.registry.member_information
+    member = member_information.get_member(membership_number)
+
+    if member is None:
+        request.session.flash(
+            "A Member with id "
+            "{} could not be found in the DB. run for the backups!".format(
+                membership_number),
+            'message_to_staff'
+        )
+        return HTTPFound(  # back to base
+            request.route_url('toolbox'))
+
+    return get_member_details(request, member)
+
+
 @view_config(renderer='templates/detail.pt',
              permission='manage',
              route_name='detail')
@@ -251,7 +309,6 @@ def member_detail(request):
     Mostly all the info about an application or membership
     in the database can be seen here.
     """
-    from decimal import Decimal as D
     logged_in = authenticated_userid(request)
     memberid = request.matchdict['memberid']
     LOG.info("member details of id %s checked by %s", memberid, logged_in)
@@ -268,23 +325,7 @@ def member_detail(request):
         return HTTPFound(  # back to base
             request.route_url('toolbox'))
 
-    # get the members invoices from the DB
-    invoices15 = Dues15Invoice.get_by_membership_no(member.membership_number)
-    invoices16 = Dues16Invoice.get_by_membership_no(member.membership_number)
-    invoices17 = Dues17Invoice.get_by_membership_no(member.membership_number)
-    shares = request.registry.share_information.get_member_shares(
-        member.membership_number)
-
-    return {
-        'today': date.today().strftime('%Y-%m-%d'),
-        'D': D,
-        'member': member,
-        'shares': shares,
-        'invoices15': invoices15,
-        'invoices16': invoices16,
-        'invoices17': invoices17,
-        # 'form': html
-    }
+    return get_member_details(request, member)
 
 
 @view_config(permission='view',